### 2022-08-12 #4675 @danielskatz
👋 @balima78 - thanks for your submission.  Due to a current shortage of editors (that we are working to resolve), I'm going to put it on our waiting list until an appropriate editor is available.

In addition, you could work on the invalid DOI that editorialbot suggests. In addition, your use of references in your .md file is not correct, as you can see in the generated PDF. Please follow the [example paper](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography). You can make changes to your .bib file, then use the command `@editorialbot check references` to check again, and use the command `@editorialbot generate pdf` when the references are right and/or when you've made changes to the paper to make a new PDF. editorialbot commands need to be the first entry in a new comment.
