+++ paper/paper.bib
@@ -48,7 +48,7 @@ year = {2018}
 }
 
 @Manual{histoc,
-  title = {histoc: Histocompatibility functions for organ transplantation},
+  title = {histoc: Histocompatibility Functions for Organ Transplantation},
   author = {Bruno Lima and Filipe Reis},
   year = {2022},
   note = {R package version 0.2.0},
