+++ man/et_mmp.Rd
@@ -4,7 +4,7 @@
 \alias{et_mmp}
 \title{Mismatch Probability (MMP) from ETKAS}
 \source{
-\url{https://www.eurotransplant.org/wp-content/uploads/2020/01/H4-Kidney.pdf}
+\url{https://www.eurotransplant.org/allocation/eurotransplant-manual/}
 }
 \usage{
 et_mmp(
