+++ DESCRIPTION
@@ -9,7 +9,7 @@ Authors@R: c(
     person(given = "TxOR",
            role = c("cph", "fnd"))
     )
-Description: This package contains functions to assess matches on organ transplantation. Histocompatibility analysis are possible with information from donors and candidates.
+Description: Provides functions to assess matches on organ transplantation. Histocompatibility analysis are possible with information from donors and candidates.
 License: MIT + file LICENSE
 BugReports: https://github.com/txopen/histoc/issues
 Encoding: UTF-8
@@ -27,7 +27,6 @@ Imports:
     utils,
     data.table,
     purrr,
-    mockr,
     tibble,
     rlang
 Depends: 
