+++ NAMESPACE
@@ -26,3 +26,6 @@ export(uk)
 export(xmatch)
 import(data.table)
 importFrom(magrittr,"%>%")
+importFrom(mockr,local_mock)
+importFrom(tibble,tribble)
+importFrom(utils,globalVariables)
