+++ paper/paper.bib
@@ -50,6 +50,7 @@
   title = {histoc: Histocompatibility Functions for Organ Transplantation},
   author = {{Lima}, B. and {Reis}, F.},
   year = {2022},
+  organization = {Transplant Open Registry},
   note = {R package version 0.2.0},
   url = {https://txopen.github.io/histoc/},
 }
