+++ R/histoc-package.R
@@ -1,9 +1,0 @@
-#' @keywords internal
-"_PACKAGE"
-
-## usethis namespace: start
-#' @importFrom mockr local_mock
-#' @importFrom tibble tribble
-#' @importFrom utils globalVariables
-## usethis namespace: end
-NULL
