+++ man/histoc-package.Rd
@@ -1,35 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/histoc-package.R
-\docType{package}
-\name{histoc-package}
-\alias{histoc}
-\alias{histoc-package}
-\title{histoc: Histocompatibility Functions for Organ Transplantation}
-\description{
-\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}
-
-Provides functions to assess matches on organ transplantation. Histocompatibility analysis are possible with information from donors and candidates.
-}
-\seealso{
-Useful links:
-\itemize{
-  \item \url{https://txopen.github.io/histoc/}
-  \item Report bugs at \url{https://github.com/txopen/histoc/issues}
-}
-
-}
-\author{
-\strong{Maintainer}: Bruno Lima \email{balima78@gmail.com} (\href{https://orcid.org/0000-0001-9090-4457}{ORCID})
-
-Authors:
-\itemize{
-  \item Filipe Reis \email{filipepintodosreis@gmail.com} (\href{https://orcid.org/0000-0002-8186-3910}{ORCID})
-}
-
-Other contributors:
-\itemize{
-  \item TxOR [copyright holder, funder]
-}
-
-}
-\keyword{internal}
