+++ inst/rascunho.R
@@ -128,4 +128,7 @@ result[!is.na(ID),][ ,
                                     mmHLA_B = mmB,
                                     mmHLA_DR = mmDR)$prob5y,
                      by = 'ID'][]
-
+##############
+et_mmp()
+et_mmHLA()
+et_dialysis()
