### 2022-08-17 b941422
fix: manuals fixed
