+++ man/et_mmp.Rd
@@ -28,8 +28,10 @@ group of waitlisted transplant candidates.}
 
 \item{abo.freq}{A data frame with ABO blood group frequencies.}
 
-\item{check.validity}{Logical to decide whether to validate input.
-#' @return An ordered data frame with a column \code{cp} (color priority),}
+\item{check.validity}{Logical to decide whether to validate input.}
+}
+\value{
+A data table with columns for HLA allele frequencies and Mismatch Probability \code{MMP} (color priority),
 }
 \description{
 Mismatch Probability (MMP) is a calculation of the probability of
