### 2022-08-17 e87a237
fix: deleted fucntion cp(), this is only used on lima()
