### 2022-08-17 e586639
fix: added poits for HLA mm
