+++ pkgdown.yml
@@ -4,7 +4,7 @@ pkgdown_sha: ~
 articles:
   cand_select: cand_select.html
   how_to: how_to.html
-last_built: 2022-08-15T18:51Z
+last_built: 2022-08-17T19:50Z
 urls:
   reference: https://txopen.github.io/histoc/reference
   article: https://txopen.github.io/histoc/articles
