### 2022-08-26 5556ca7
feat: checker functions for donors' files
