### 2022-08-31 1ecc9c5
Added hed calculation functions as is
