+++ DESCRIPTION
@@ -27,7 +27,8 @@ Imports:
     purrr,
     rlang,
     mockr,
-    tibble
+    tibble,
+    seqinr
 Depends: 
     R (>= 4.1.0)
 LazyData: true
