### 2022-08-31 90de831
Simplified calculate distance usage
