### 2022-09-03 61c621e
feat: added a new funtion: cHED()
