+++ man/cHED.Rd
@@ -1,0 +1,24 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/hed.R
+\name{cHED}
+\alias{cHED}
+\title{calculated HLA Evolutionary Divergence (HED) score.}
+\usage{
+cHED(hla1, hla2)
+}
+\arguments{
+\item{hla1}{first HLA allele.}
+
+\item{hla2}{second HLA allele.}
+}
+\value{
+A numerical values corresponding to the divergence between alleles.
+}
+\description{
+givem two HLA classe I alleles calculates HLA Evolutionary Divergence, i.e,
+divergenced between allele sequences using Grantham distance metric
+}
+\examples{
+cHED(hla1 = "A*02:653",
+ hla2 = "A*02:654")
+}
