+++ man/calculate_evolutionary_distance.Rd
@@ -1,0 +1,37 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/utility.R
+\name{calculate_evolutionary_distance}
+\alias{calculate_evolutionary_distance}
+\title{Given two hla sequences, calculate the HLA evolutionary distances between them.}
+\usage{
+calculate_evolutionary_distance(
+  hla1,
+  hla2,
+  sequence.file.path,
+  distance.file.path
+)
+}
+\arguments{
+\item{hla1}{First sequence identifier.}
+
+\item{hla2}{Second sequence identifier.}
+
+\item{sequence.file.path}{Path of a fasta file containing nucleic acid sequences.
+Creates a labeled list with the identifier-sequence pairs.}
+
+\item{distance.file.path}{Path of a txt matrix with the nucleic acid distances.
+Creates a data.table with the same information.}
+}
+\value{
+The distance between the two hla sequences as a number.
+}
+\description{
+Given two hla sequences, calculate the HLA evolutionary distances between them.
+}
+\examples{
+calculate_evolutionary_distance(hla1 = "A*02:653", 
+ hla2 = "A*02:654",
+ sequence = "./inst/database/ABC_prot.fa", 
+ distance = "./inst/database/grantham_matrix.txt"
+)
+}
