+++ NAMESPACE
@@ -5,6 +5,7 @@ export(abo)
 export(abo_uk)
 export(age_diff)
 export(b_blood_penalization)
+export(calculate_evolutionary_distance)
 export(cp)
 export(donor_recipient_pairs)
 export(et)
