### 2022-09-06 8715759
feat: added unitary tests for cHED(); new version
