### 2022-09-16 f596394
feat: cHED() now allows to compute HLA classe II HED
