+++ revdep/README.md
@@ -1,0 +1,20 @@
+These are *partial* results!
+
+# Platform
+
+|field    |value                                    |
+|:--------|:----------------------------------------|
+|version  |R version 4.1.2 (2021-11-01)             |
+|os       |Windows 10 x64 (build 19044)             |
+|system   |x86_64, mingw32                          |
+|ui       |RStudio                                  |
+|language |(EN)                                     |
+|collate  |Portuguese_Portugal.1252                 |
+|ctype    |Portuguese_Portugal.1252                 |
+|tz       |Europe/London                            |
+|date     |2022-07-21                               |
+|rstudio  |2022.02.0+443 Prairie Trillium (desktop) |
+|pandoc   |NA                                       |
+
+# Dependencies
+
