+++ DESCRIPTION
@@ -12,7 +12,7 @@ License: MIT + file LICENSE
 BugReports: https://github.com/txopen/histoc/issues
 Encoding: UTF-8
 Roxygen: list(markdown = TRUE)
-RoxygenNote: 7.2.0
+RoxygenNote: 7.2.1
 Suggests:
     covr,
     knitr,
