+++ R/globals.R
@@ -21,5 +21,5 @@ utils::globalVariables(c("valid_blood_groups", "valid_blood_groups",
                          "MMP1","level",
                          "donors","dABO","dA","dB","dDR","donor.age",
                          "algorithm", "recipient.dialysis", "donors.uk", "check.validity",
-                         "seqs","dst"
+                         "seqs","dst", "seqs2"
                          ))
