+++ R/data.R
@@ -208,7 +208,7 @@
 #' HLA-A, -B, -C amino-acid sequences in FASTA format.
 #'
 #' The protein sequence of exons 2 and 3 of each allele of HLA classe I genotype;
-#' these sequences correspond to the peptide-binding domains.
+#' these sequences correspond to the peptide-binding domains. Availabble from: https://www.ebi.ac.uk/ipd/imgt/hla/
 #'
 #' @format list in FASTA format representing either nucleotide sequences or amino acid (protein) sequences, in which nucleotides or amino acids are represented using single-letter codes:
 #' \describe{
@@ -218,7 +218,7 @@
 #'   \item{class}{SeqFastadna}
 #'   ...
 #' }
-#' @source \url{https://www.ebi.ac.uk/ipd/imgt/hla/}
+#' @source \url{https://github.com/sunhuaibo/HLA-HED/blob/main/database/ABC_prot.fa}
 "seqs"
 
 #' HLA-DRB1, -DQB1, -DPB1 amino-acid sequences in FASTA format.
@@ -234,8 +234,8 @@
 #'   \item{class}{SeqFastadna}
 #'   ...
 #' }
-#' @source \url{https://www.ebi.ac.uk/ipd/imgt/hla/}
-"seqs"
+#' @source \url{https://sourceforge.net/projects/granthamdist/files/HLA_ClassII_CWDonly.fas/download}
+"seqs2"
 
 #' Grantham distance metric.
 #'
