### 2022-09-19 1140bbb
fix: added unit tests for cHED() and corrected for class II
