+++ inst/rascunho.R
@@ -107,3 +107,8 @@ s5
 
 ###############################
 
+cHED(hla1 = 'DRB10701', hla2 = 'DRB11302')
+
+cHED('DRB11104','RB11501')
+
+names(seqs2)
