### 2022-09-19 87c20a6
fix: cHED() classe II allele normal nomenclature acepted
