### 2022-09-28 #4675 @aj2duncan
> 👋 - @aj2duncan Would you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Hi @crvernon , I should be able to help with a review on this. Thanks for the invite.
