### 2022-09-28 #4807 @editorialbot
[REVIEW]: {histoc}  histocompatibility functions for organ transplantation
**Submitting author:** @balima78 (<a href="http://orcid.org/0000-0001-9090-4457">Bruno A Lima</a>)
**Repository:** https://github.com/txopen/histoc
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.1
