### 2022-09-28 #4807 @crvernon
👋 @balima78,  @turgeonmaxime, @aj2duncan - This is the review thread for the paper. All of our communications will happen here from now on.
