### 2022-10-12 #4807 @crvernon
📣 Mid-week rally!

👋 @turgeonmaxime and  @aj2duncan  - Can you provide a status update to how things are going on your side of things?  Thanks!

Keep up the good work!
