### 2022-10-13 #4807 @turgeonmaxime
@balima78 I've read the manuscript a couple of times. I think the statement of need needs to be expanded. For example, I can't find the answer to the following questions:

- Who is the intended audience for the package? Physicians? Researchers in kidney transplant allocation policy? Biostatisticians? Health systems administrator?
- What is the current state of software solutions for this problem?
- How does `histoc` compare to current software?
