### 2022-10-17 #4807 @aj2duncan
Hi @balima78,

I agree with @turgeonmaxime that the statement of need in the manuscript could have some additional info added. I would also suggest that you need to add some more detail to the guidelines for contribution.

Thanks.
