### 2022-10-29 0c887b0
KPSAM rephrased
