+++ R/data.R
@@ -210,7 +210,7 @@
 #' The protein sequence of exons 2 and 3 of each allele of HLA classe I genotype;
 #' these sequences correspond to the peptide-binding domains. Availabble from: https://www.ebi.ac.uk/ipd/imgt/hla/
 #'
-#' @format list in FASTA format representing either nucleotide sequences or amino acid (protein) sequences, in which nucleotides or amino acids are represented using single-letter codes:
+#' @format list in FASTA format representing either nucleotide sequences or amino-acid (protein) sequences, in which nucleotides or amino-acids are represented using single-letter codes:
 #' \describe{
 #'   \item{X*}{HLA-X allele}
 #'   \item{name}{allele name}
@@ -223,10 +223,10 @@
 
 #' HLA-DRB1, -DQB1 amino-acid sequences in FASTA format.
 #'
-#' The protein sequence of exon 2 of each allele of HLA classe II genotype;
+#' The protein sequence of exon 2 of each allele of HLA class II genotype;
 #' these sequences correspond to the peptide-binding domains.
 #'
-#' @format list in FASTA format representing either nucleotide sequences or amino acid (protein) sequences, in which nucleotides or amino acids are represented using single-letter codes:
+#' @format list in FASTA format representing either nucleotide sequences or amino-acid (protein) sequences, in which nucleotides or amino-acids are represented using single-letter codes:
 #' \describe{
 #'   \item{X*}{HLA-X allele}
 #'   \item{name}{allele name}
@@ -239,14 +239,14 @@
 
 #' Grantham distance metric.
 #'
-#' The Grantham score attempts to predict the distance between two amino acids, in an evolutionary sense.
+#' The Grantham score attempts to predict the distance between two amino-acids, in an evolutionary sense.
 #' A lower Grantham score reflects less evolutionary distance.
 #' A higher Grantham score reflects a greater evolutionary distance.
 #'
 #' @format a 20x20 matrix
 #' \describe{
-#'   \item{rows}{amino acid}
-#'   \item{columns}{amino acid}
+#'   \item{rows}{amino-acid}
+#'   \item{columns}{amino-acid}
 #' }
 #' @source \url{https://www.science.org/doi/10.1126/science.185.4154.862}
 "dst"
