+++ man/seqs2.Rd
@@ -5,7 +5,7 @@
 \alias{seqs2}
 \title{HLA-DRB1, -DQB1 amino-acid sequences in FASTA format.}
 \format{
-list in FASTA format representing either nucleotide sequences or amino acid (protein) sequences, in which nucleotides or amino acids are represented using single-letter codes:
+list in FASTA format representing either nucleotide sequences or amino-acid (protein) sequences, in which nucleotides or amino-acids are represented using single-letter codes:
 \describe{
 \item{X*}{HLA-X allele}
 \item{name}{allele name}
@@ -21,7 +21,7 @@ list in FASTA format representing either nucleotide sequences or amino acid (pro
 seqs2
 }
 \description{
-The protein sequence of exon 2 of each allele of HLA classe II genotype;
+The protein sequence of exon 2 of each allele of HLA class II genotype;
 these sequences correspond to the peptide-binding domains.
 }
 \keyword{datasets}
