+++ R/data.R
@@ -207,8 +207,8 @@
 
 #' HLA-A, -B, -C amino-acid sequences in FASTA format.
 #'
-#' The protein sequence of exons 2 and 3 of each allele of HLA classe I genotype;
-#' these sequences correspond to the peptide-binding domains. Availabble from: https://www.ebi.ac.uk/ipd/imgt/hla/
+#' The protein sequence of exons 2 and 3 of each allele of HLA class I genotype;
+#' these sequences correspond to the peptide-binding domains. Available from: https://www.ebi.ac.uk/ipd/imgt/hla/
 #'
 #' @format list in FASTA format representing either nucleotide sequences or amino acid (protein) sequences, in which nucleotides or amino acids are represented using single-letter codes:
 #' \describe{
@@ -223,7 +223,7 @@
 
 #' HLA-DRB1, -DQB1 amino-acid sequences in FASTA format.
 #'
-#' The protein sequence of exon 2 of each allele of HLA classe II genotype;
+#' The protein sequence of exon 2 of each allele of HLA class II genotype;
 #' these sequences correspond to the peptide-binding domains.
 #'
 #' @format list in FASTA format representing either nucleotide sequences or amino acid (protein) sequences, in which nucleotides or amino acids are represented using single-letter codes:
