### 2022-10-31 #4807 @turgeonmaxime
@balima78 Thanks for adding more details about the state of available software.

I did notice a few typos:

- Line 7: Should be "as equitable as possible", instead of "equitably".
- Line 36: "a vignette is available", instead of "it is available a vignette".
- Line 41: "makes available", instead of "make available".
- Subsection on _Portugese Model_: The numbered list didn't render properly. I think it's because there should be an empty line between the end of the paragraph and the beginning of the list.
- Line 57: Should be "hypersensitized", instead of "hipersensitized".
- Line 99: Should be "dialysis", instead of "dialisys".
- Line 112: There is an extra bracket ] that should be remove.
