### 2022-11-02 #4807 @aj2duncan
Hi @balima78,

Thanks very much for those changes. I think you just need to make a couple of additions to your documentation/github and everything is complete. At the moment the paper has both a Statement of Need and details on Bug Reports and How to Contribute, I would suggest you add these to Github and/or [your documentation](https://txopen.github.io/histoc/) as you see fit - just in case someone doesn't find the paper.

By the way, one of your tests is skipped when running `testthat::test_local()` as it is empty. Otherwise, this is a well written, tested and document package.
