### 2022-11-03 #4807 @balima78
thank you @aj2duncan, I added the 'Bug report' to the package's [README](https://github.com/txopen/histoc) and website, as suggested.
