+++ paper/paper.bib
@@ -89,7 +89,7 @@
 }
 
 @misc{UK,
-      author = {National Health Service Blood and Transplant},
+      author = {{National Health Service Blood and Transplant}},
       title = {{UK Transplant Registry, Organ Donation and Transplantation}},
       url = {http://odt.nhs.uk/uk-transplant-registry/data/},
       year = {2017}
@@ -105,8 +105,15 @@
 
 @misc{srtr,
       title = {Kidney-Pancreas Simulated Allocation Model (KPSAM)},
-      author = {Scientific Registry of Transplant Recipients},
+      author = {{Scientific Registry of Transplant Recipients}},
       year = {2015},
       url = {https://srtr.org/requesting-srtr-data/simulated-allocation-models/},
       note = {R package version 0.2.0},
 }
+
+@misc{PT,
+      title = {Ordinance nº 6537/2007},
+      author = {{Ministry of Health}},
+      year = {2007},
+      url = {http://www.ipst.pt/files/TRANSPLANTACAO/DOACAOETRANSPLANTACAO/Despacho_6537_2007.pdf}
+      }
