+++ paper/paper.tex
@@ -97,7 +97,7 @@ allocation systems}
 \section{Summary}\label{summary}}
 
 The distribution of a scarce commodity such as deceased donor's kidneys
-for transplantation should be as equitably as possible. Different
+for transplantation should be as equitable as possible. Different
 countries try to implement kidney allocation systems (KAS) in
 transplantation that balance principles of justice and utility in the
 distribution of such scarce resource
@@ -119,58 +119,61 @@ in transplantation.
 \hypertarget{statement-of-need}{%
 \section{Statement of need}\label{statement-of-need}}
 
-\{histoc\} (\protect\hyperlink{ref-histoc}{Lima and Reis 2022}) is an R
-(\protect\hyperlink{ref-R}{R Core Team 2021}) package that assembles
-tools for histocompatibility testing in the context of kidney
-transplantation. The package's main functions allow simulating several
-KAS on the distribution of deceased donors' grafts for transplantation.
-Moreover, it is possible to redefine arguments for each one of the KAS
-as a way to test different approaches.
+\texttt{histoc} is an R (\protect\hyperlink{ref-R}{R Core Team 2021})
+package that assembles tools for histocompatibility testing in the
+context of kidney transplantation. The package's main functions allow
+simulating several KAS on the distribution of deceased donors' grafts
+for transplantation. Moreover, it is possible to redefine arguments for
+each one of the KAS as a way to test different approaches.
 
 Currently, it is possible to simulate allocation rules implemented in
-Portugal (PT model), in countries within Eurotransplant (ET model)
-(\protect\hyperlink{ref-ET}{EuroTransplant 2020}), in the United Kingdom
-(UK model) (\protect\hyperlink{ref-UK}{Blood and Transplant 2017}), and
-a system suggested by \protect\hyperlink{ref-Lima:2013}{Lima, Mendes,
-and Alves} (\protect\hyperlink{ref-Lima:2013}{2013}) (Lima's model).
+Portugal {[}PT model; (\protect\hyperlink{ref-PT}{\textbf{PT?}}){]}, in
+countries within Eurotransplant {[}ET model;
+\protect\hyperlink{ref-ET}{EuroTransplant}
+(\protect\hyperlink{ref-ET}{2020}){]}, in the United Kingdom {[}UK
+model; \protect\hyperlink{ref-UK}{Blood and Transplant}
+(\protect\hyperlink{ref-UK}{2017}){]}, and a system suggested by
+\protect\hyperlink{ref-Lima:2013}{Lima, Mendes, and Alves}
+(\protect\hyperlink{ref-Lima:2013}{2013}) (Lima's model).
 
 Each one of these models have as arguments a data frame with transplant
 candidates' clinical and demographic characteristics, a data frame with
-candidates' HLA antibodies and potential donor's information.
+candidates' Human Leukocyte Antigens (HLA) antibodies and potential
+donor's information. By default two candidates are selected for each
+donor, although we can define the number of candidates to be selected.
 
 For all the models a virtual crossmatch between the donor and transplant
 candidates is performed (\texttt{xmatch()}). And, only those candidates
 with a negative crossmatch and ABO compatible can opt to a donor's
 kidney.
 
-Results are presented as \{data.table\}
+Results are presented as \texttt{data.table}
 (\protect\hyperlink{ref-data.table}{Dowle and Srinivasan 2021}) objects
 due to its high computation performance.
 
-To get started a vignette describes
+To get started, a vignette is available that describes
 \href{https://txopen.github.io/histoc/articles/how_to.html}{how to use}
-each of the algorithms.
-
-By default 2 candidates are selected for each donor, although we can
[…]
