+++ paper/paper.md
@@ -6,7 +6,9 @@ tags:
 - kidney
 - allocation
 date: "24 July 2022"
-output: pdf_document
+output:
+  html_document:
+    df_print: paged
 authors:
 - name: Bruno A Lima
   orcid: 0000-0001-9090-4457
@@ -33,27 +35,27 @@ The goal of this package is to aid the evaluation and assessment of KAS in trans
 
 # Statement of need
 
-{histoc} [@histoc] is an R [@R] package that assembles tools for histocompatibility testing in the context of kidney transplantation. The package's main functions allow simulating several KAS on the distribution of deceased donors’ grafts for transplantation. Moreover, it is possible to redefine arguments for each one of the KAS as a way to test different approaches.
+`histoc` is an R [@R] package that assembles tools for histocompatibility testing in the context of kidney transplantation. The package's main functions allow simulating several KAS on the distribution of deceased donors’ grafts for transplantation. Moreover, it is possible to redefine arguments for each one of the KAS as a way to test different approaches.
 
-Currently, it is possible to simulate allocation rules implemented in Portugal (PT model), in countries within Eurotransplant (ET model) [@ET], in the United Kingdom (UK model) [@UK], and a system suggested by @Lima:2013 (Lima’s model).
+Currently, it is possible to simulate allocation rules implemented in Portugal [PT model, @PT], in countries within Eurotransplant [ET model, @ET], in the United Kingdom [UK model, @UK], and a system suggested by @Lima:2013 (Lima’s model).
 
-Each one of these models have as arguments a data frame with transplant candidates’ clinical and demographic characteristics, a data frame with candidates’ HLA antibodies and potential donor's information. By default 2 candidates are selected for each donor, although we can define the number of candidates to be selected.
+Each one of these models have as arguments a data frame with transplant candidates’ clinical and demographic characteristics, a data frame with candidates’ Human Leukocyte Antigens (HLA) antibodies and potential donor's information. By default two candidates are selected for each donor, although we can define the number of candidates to be selected.
 
 For all the models a virtual crossmatch between the donor and transplant candidates is performed (`xmatch()`). And, only those candidates with a negative crossmatch and ABO compatible can opt to a donor’s kidney.
 
-Results are presented as {data.table} [@data.table] objects due to its high computation performance.
+Results are presented as `data.table` [@data.table] objects due to its high computation performance.
 
 To get started, a vignette is available that describes [how to use](https://txopen.github.io/histoc/articles/how_to.html) each one of the algorithms.
 
-New kidney allocation systems should be assessed using simulations that, to the greatest extent possible, can predict outcomes. {histoc} is designed mainly for researchers working on organ transplantation, assisting with data-driven decision making for the establishment of allocation policies.
+New kidney allocation systems should be assessed using simulations that, to the greatest extent possible, can predict outcomes. `histoc` is designed mainly for researchers working on organ transplantation, assisting with data-driven decision making for the establishment of allocation policies.
 
-While the R package {transplantr} [@transplantr] makes available a set of functions for audit and clinical research in transplantation, the package presented here enables the simulation of various sets of rules by adjusting relevant allocation parameters. Additionally, the Kidney-Pancreas Simulated Allocation Model (KPSAM) [@srtr] is a proprietary software that the Scientific Registry of Transplant Recipients makes available to support studies on alternative allocations policies in transplantation. In contrast, {histoc}, coupled with being open source, needs less data to run in comparison to KPSAM software. Likewise, it can be used as a preliminary technique for developing new hypotheses that can then be tested on KPSAM.
[…]
