### 2022-11-17 #4807 @balima78
@crvernon, thanks for your insigths.
I corrected all itens but the one about ABO, because it is not an acronym (https://www.ncbi.nlm.nih.gov/books/NBK2267/)
