### 2022-11-18 #4807 @crvernon
@balima78 after we confirm that the above compiles in a way that suits you and the changes are brought into your `main` branch, we can move on to setting up the archive for a new release:

- We want to make sure the archival has the correct metadata that JOSS requires. This includes a title that matches the paper title and a correct author list.

So here is what we have left to do after the most recent paper changes have been made:

- [ ] Conduct a GitHub release of the current reviewed version of the software you now have on the main and archive the reviewed software in Zenodo or a similar service (e.g., figshare, an institutional repository)
- [ ] List the DOI of the archived version here
