### 2022-12-01 #4807 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
